+++ platform/linux-dpdk/odp_buffer.c
@@ -20,7 +20,7 @@ void *odp_buffer_addr(odp_buffer_t buf)
 }
 
 
-size_t odp_buffer_size(odp_buffer_t buf)
+uint32_t odp_buffer_size(odp_buffer_t buf)
 {
 	odp_buffer_hdr_t *hdr = odp_buf_to_hdr(buf);
 	struct rte_mbuf *mbuf = (struct rte_mbuf *)hdr;
